refactor: simplify imports from icalendar.cal package
