+++ news/1707.chore.1
@@ -1,0 +1,2 @@
+Simplified type checking imports in :mod:`icalendar.cal.alarm`, `icalendar.cal.availability`, `icalendar.cal.event`, `icalendar.cal.free_busy`, `icalendar.cal.journal`, `icalendar.cal.lazy`, `icalendar.cal.timezone`, and `icalendar.cal.todo` to import directly from the package instead of submodules.
+@Priyanshu-pulak
