+++ src/icalendar/cal/todo.py
@@ -51,11 +51,14 @@ from icalendar.cal.examples import get_example
 if TYPE_CHECKING:
     from collections.abc import Iterable, Sequence
 
-    from icalendar.alarms import Alarms
-    from icalendar.compatibility import Self
-    from icalendar.enums import CLASS, STATUS
-    from icalendar.prop import vCalAddress
-    from icalendar.prop.conference import Conference
+    from icalendar import (
+        CLASS,
+        STATUS,
+        Alarms,
+        Self,
+        vCalAddress,
+    )
+    from icalendar.prop import Conference
 
 
 class Todo(Component):
